// types/about.ts
export interface About {
  id: string;
  title_en: string;
  title_ru: string;
  title_uz: string;
  description_en: string;
  description_ru: string;
  description_uz: string;
  location_en: string;
  location_ru: string;
  location_uz: string;
  availability_en: string;
  availability_ru: string;
  availability_uz: string;
  education_en: string;
  education_ru: string;
  education_uz: string;
  what_i_do_en: string;
  what_i_do_ru: string;
  what_i_do_uz: string;
  service_1_en: string;
  service_1_ru: string;
  service_1_uz: string;
  service_2_en: string;
  service_2_ru: string;
  service_2_uz: string;
  service_3_en: string;
  service_3_ru: string;
  service_3_uz: string;
  service_4_en: string;
  service_4_ru: string;
  service_4_uz: string;
  image_url: string | null;
  is_published: boolean;
  created_at: string;
  updated_at: string;
}

export interface UpdateAboutDTO {
  id: string;
  title_en?: string;
  title_ru?: string;
  title_uz?: string;
  description_en?: string;
  description_ru?: string;
  description_uz?: string;
  location_en?: string;
  location_ru?: string;
  location_uz?: string;
  availability_en?: string;
  availability_ru?: string;
  availability_uz?: string;
  education_en?: string;
  education_ru?: string;
  education_uz?: string;
  what_i_do_en?: string;
  what_i_do_ru?: string;
  what_i_do_uz?: string;
  service_1_en?: string;
  service_1_ru?: string;
  service_1_uz?: string;
  service_2_en?: string;
  service_2_ru?: string;
  service_2_uz?: string;
  service_3_en?: string;
  service_3_ru?: string;
  service_3_uz?: string;
  service_4_en?: string;
  service_4_ru?: string;
  service_4_uz?: string;
  image_url?: string | null;
  is_published?: boolean;
}